Refactor: 버튼, 링크 객체 상수화
