+++ src/components/login/LoginButton.tsx
@@ -4,24 +4,12 @@ import Image from "next/image";
 
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
-
 import "@/components/login/LoginButton.scss";
 
-type ButtonInfo = Array<{
-  name: string;
-  text: string;
-  img: string;
-}>;
-const BUTTON_INFO: ButtonInfo = [
-  { name: "naver", text: "네이버 로그인", img: "/images/naver.svg" },
-  { name: "kakao", text: "카카오톡 로그인", img: "/images/kakao.svg" },
-  { name: "google", text: "구글 로그인", img: "/images/google.svg" },
-];
+import { BUTTON_INFO, LOGIN_LINKS } from "@/lib/util/login/loginButton";
 
 function LoginButton() {
   const router = useRouter();
-  const link = `https://kauth.kakao.com/oauth/authorize?client_id=${process.env.NEXT_PUBLIC_KAKAO_REST_API_KEY}&redirect_uri=${process.env.NEXT_PUBLIC_KAKAO_REDIRECT_URI}&response_type=code
-  `;
 
   useEffect(() => {
     const authToken = localStorage.getItem("authToken");
@@ -30,9 +18,14 @@ function LoginButton() {
     }
   }, [router]);
 
-  const onClickButton = async () => {
+  const onClickButton = (event: React.MouseEvent<HTMLButtonElement>) => {
+    const name = event.currentTarget.getAttribute("data-name");
+    if (!name || !LOGIN_LINKS[name]) {
+      console.error("로그인 링크를 찾을 수 없음");
+      return;
+    }
     try {
-      window.location.href = link;
+      window.location.href = LOGIN_LINKS[name];
     } catch (error) {
       console.error("로그인 실패:", error);
     }
@@ -44,6 +37,7 @@ function LoginButton() {
         <button
           onClick={onClickButton}
           key={name}
+          data-name={name}
           className={classNames(
             "buttons-block__button",
             `buttons-block__button--${name}`,
